+++ lib/elasticsearch_record/relation/calculation_methods.rb
@@ -33,9 +33,9 @@ module ElasticsearchRecord
           # since total will be limited to 10000 results, we need to resolve the real values by a custom query.
           # This query is called through +#select_count+.
           #
-          # HINT: :__claim__ directly interacts with the query-object and sets a 'terminate_after' argument
-          # (see @ Arel::Collectors::ElasticsearchQuery#assign)
-          arel = spawn.unscope!(:offset, :limit, :order, :configure, :aggs).configure!(:__claim__, argument: { terminate_after: limit_value }).arel
+          # HINT: +:__query__+ directly interacts with the query-object and sets the 'terminate_after' argument
+          # see @ ElasticsearchRecord::Query#arguments & Arel::Collectors::ElasticsearchQuery#assign
+          arel = spawn.unscope!(:offset, :limit, :order, :configure, :aggs).configure!(:__query__, argument: { terminate_after: limit_value }).arel
           klass.connection.select_count(arel, "#{klass.name} Count")
         else
           # since total will be limited to 10000 results, we need to resolve the real values by a custom query.